move check-enforcements back into test/support
introduced #25384
fixes #25454
