+++ generators/cypress/generator.spec.ts
@@ -21,8 +21,8 @@ import { fileURLToPath } from 'url';
 import { before, it, describe, after, expect } from 'esmocha';
 import lodash from 'lodash';
 import { clientFrameworkTypes, testFrameworkTypes } from '../../jdl/jhipster/index.js';
-import { fromMatrix, extendMatrix, AuthenticationTypeMatrix, checkEnforcements, defaultHelpers as helpers } from '../../testing/index.js';
-import { shouldSupportFeatures, testBlueprintSupport } from '../../test/support/tests.js';
+import { fromMatrix, extendMatrix, AuthenticationTypeMatrix, defaultHelpers as helpers } from '../../testing/index.js';
+import { shouldSupportFeatures, testBlueprintSupport, checkEnforcements } from '../../test/support/index.js';
 import Generator from './generator.js';
 import { GENERATOR_CYPRESS } from '../generator-list.js';
