+++ generators/feign-client/generator.spec.ts
@@ -1,8 +1,8 @@
 import { basename, dirname } from 'path';
 import { fileURLToPath } from 'url';
 import { before, it, describe, expect } from 'esmocha';
-import { checkEnforcements, defaultHelpers as helpers, runResult } from '../../testing/index.js';
-import { shouldSupportFeatures, testBlueprintSupport } from '../../test/support/tests.js';
+import { defaultHelpers as helpers, runResult } from '../../testing/index.js';
+import { shouldSupportFeatures, testBlueprintSupport, checkEnforcements } from '../../test/support/index.js';
 import Generator from '../server/index.js';
 
 import { GENERATOR_FEIGN_CLIENT } from '../generator-list.js';
