+++ generators/spring-boot/generator.spec.ts
@@ -2,8 +2,8 @@ import { basename, dirname } from 'path';
 import { fileURLToPath } from 'url';
 import { before, it, describe, expect } from 'esmocha';
 import { snakeCase } from 'lodash-es';
-import { checkEnforcements, defaultHelpers as helpers, runResult } from '../../testing/index.js';
-import { shouldSupportFeatures, testBlueprintSupport } from '../../test/support/tests.js';
+import { defaultHelpers as helpers, runResult } from '../../testing/index.js';
+import { shouldSupportFeatures, testBlueprintSupport, checkEnforcements } from '../../test/support/index.js';
 import Generator from '../server/index.js';
 
 import { mockedGenerators } from '../server/__test-support/index.js';
