+++ test/support/index.ts
@@ -1,0 +1,2 @@
+export { default as checkEnforcements } from './check-enforcements.js';
+export * from './tests.js';
