+++ testing/index.ts
@@ -4,7 +4,6 @@ export * from './helpers.js';
 // test matrix
 export * from './support/application-samples.js';
 export * from './support/client-samples.js';
-export { default as checkEnforcements } from './support/check-enforcements.js';
 export * from './support/entity-samples.js';
 export * from './support/matrix-utils.js';
 export * from './support/matcher.js';
